Generalize some commonly-used testing code
